新Firebase Admin SDKに対応: tabulate系
